keos_project5::ffs::access_control

Struct TrackedInode

pub struct TrackedInode(Arc<RwLock<Inode>>, Weak<FastFileSystemInner>);

A reference-counted, thread-safe wrapper around an in-memory Inode, enabling synchronized read and transactional write access.

This type provides read access through a shared guard and write access via a transactional context, ensuring consistency between the in-memory and on-disk representations of the inode.

Tuple Fields
0: Arc<RwLock<Inode>>1: Weak<FastFileSystemInner>
Implementations

impl TrackedInode

pub fn new(inner: Arc<RwLock<Inode>>, ffs: Weak<FastFileSystemInner>) -> Self

Create a new TrackedInode reference.

pub fn read(&self) -> TrackedInodeReadGuard<'_>

Acquires a shared read lock on the in-memory inode.

Returns

A TrackedInodeReadGuard which provides read-only access to the current state of the inode.

Use Case

Use this when you need to inspect an inode without modifying it.

pub fn write_with<R>( &self, tx: &RunningTransaction<'_>, f: impl FnOnce(TrackedInodeWriteGuard<'_, '_, '_, '_>) -> Result<R, KernelError>, ) -> Result<R, KernelError>

Acquires an exclusive write lock on both the in-memory inode and the corresponding on-disk inode for transactional modification.

You must submit the changes by calling the submit method.

Arguments

tx: A reference to the current RunningTransaction used to track and commit filesystem changes.

f: A closure that performs modifications using the provided TrackedInodeWriteGuard, which contains both in-memory and on-disk representations of the inode.

Returns

Ok(R): If the closure f returns successfully.

Err(KernelError): If an error occurs while resolving the inode layout or during execution of f.

Use Case

Use this when updating inode state (e.g., growing a file, updating metadata) and ensuring consistency between memory and disk through the transaction.

Example

tracked_inode.write_with(tx, |mut guard| {
    guard.mem_layout.size += 1;
    guard.disk_layout[guard.index].size = guard.mem_layout.size;
    guard.submit();
    Ok(())
})?;