keos_project5::ffs::access_control

Struct TrackedInodeWriteGuard

pub struct TrackedInodeWriteGuard<'a, 'b, 'c, 'd> {
    mem_layout: RwLockWriteGuard<'a, Inode>,
    disk_layout: BlockPointsToWriteGuard<'b, 'c, 'd, InodeArray>,
    index: usize,
}

A guard that provides mutable access to the in-memory within a transactional context.

This guard is constructed internally by TrackedInode::write_with and ensures that any modifications are performed consistently across both memory and disk. The changes must be explicitly committed using the associated transaction.

Panics

If dropped without submitting the change to the transaction, this guard will panic to ensure no silent inconsistency between in-memory and on-disk state.

Example

tracked_inode.write_with(tx, |mut guard| {
    guard.mem_layout.size += 1;
    guard.disk_layout[guard.index].size = guard.mem_layout.size;
    guard.submit();
    Ok(())
})?;

Fields
mem_layout: RwLockWriteGuard<'a, Inode>

Write guard protecting the in-memory inode.

disk_layout: BlockPointsToWriteGuard<'b, 'c, 'd, InodeArray>

Write guard protecting the on-disk inode array.

index: usize

Index to the inode.

Implementations

impl TrackedInodeWriteGuard<'_, '_, '_, '_>

pub fn submit(self)

Submits the modified metadata block to the RunningTransaction.

This function marks the block as dirty and ensures it will be written to disk as part of the journal. After calling submit, the guard is consumed.

fn do_submit(self)