keos_project5::ffs::disk_layout

Struct DirectoryBlockEntry

#[repr(C)]
pub struct DirectoryBlockEntry {
    pub inode: Option<InodeNumber>,
    pub name_len: u8,
    pub name: [u8; 251],
}

Represent a single directory entry within a directory block.

Each entry stores metadata for necessary to locate a file or subdirectory.

Fields
inode: Option<InodeNumber>

The inode associated with this directory entry.

Some(inode): a valid file or directory.

None: an unused or deleted entry.

name_len: u8

The length of the file or directory name stored in name. This indicates how many bytes in the name array are valid.

name: [u8; 251]

The name of the file or directory. Only the first name_len bytes are meaningful.

Implementations

impl DirectoryBlockEntry

pub fn from_ino_name(ino: InodeNumber, name: &str) -> Option<Self>

Constructs a new directory entry from an inode number and name.

Returns None if the name is too long to fit in the directory entry.

Arguments

ino: The inode number associated with the entry.

name: The name of the file or directory.

Returns

Some(Self): A valid directory entry.

None: If the name is too long to fit.

pub fn name(&self) -> Option<&str>

Returns the name of the directory entry as a string slice.

Returns

Some(&str): If the name is valid UTF-8.

None: If the name contains invalid UTF-8 bytes.