keos_project5::ffs::disk_layout

Struct IndirectBlock

#[repr(C)]
pub struct IndirectBlock {
    lbas: [Option<LogicalBlockAddress>; 512],
}

Represents an indirect block in the filesystem.

An indirect block is used to extend the number of data blocks that an inode can reference. Instead of storing data directly, it contains a list of logical block addresses (LBAs), each pointing to a separate data block on disk.

This structure allows files to grow beyond the direct block limit imposed by the inode structure.

Usage

Typically used as part of indirect, or double-indirectblock addressing schemes to support large file sizes.

Fields
lbas: [Option<LogicalBlockAddress>; 512]