keos_project5::ffs::disk_layout

Struct InodeArray

#[repr(C, packed(1))]
pub struct InodeArray {
    inodes: [Inode; 16],
}

Represents a fixed-size array of inodes loaded from a block.

Each InodeArray holds a collection of Inode structures that are packed into a single 4KB block. The number of inodes it contains is computed based on the size of the Inode type.

Fields
inodes: [Inode; 16]