keos_project5::ffs::disk_layout

Struct SuperBlock

#[repr(C)]
pub struct SuperBlock {
    pub magic: [u8; 8],
    pub block_count: u64,
    pub block_count_inused: u64,
    pub inode_count: u64,
    pub inode_count_inused: u64,
    pub has_journal: u64,
    pub _pad: [u8; 4048],
}

On-disk representation of the superblock of the Fast File System (FFS).

The superblock contains essential metadata about the filesystem, including the total number of blocks and inodes, as well as the size of the journal.

Fields
magic: [u8; 8]

File system magic: โKeOSFFS\0โ.

block_count: u64

Total number of blocks in the filesystem.

block_count_inused: u64

In-used count of blocks in the filesystem.

inode_count: u64

Total number of inodes in the filesystem.

inode_count_inused: u64

In-used count of inodes in the filesystem.

has_journal: u64

A indicator that this filesystem have journaling feature.

_pad: [u8; 4048]

Padding to align to Block size.

Implementations

impl SuperBlock

pub fn from_disk(disk: &Disk) -> Result<BlockPointsTo<Self>, KernelError>

Loads the superblock structure from disk.

This function reads the first 8 sectors (4096 bytes) from the disk. It is the first step when mounting a file system, as the superblock contains metadata such as layout information, and journals.

Parameters

disk: A reference to the block device implementing the Disk trait.

Returns

Ok(Box<SuperBlock>): If the superblock is successfully read.

Err(KernelError): If any sector read fails.