keos_project5::ffs::fs_objects

Struct Directory

pub struct Directory {
    pub ffs: Weak<FastFileSystemInner>,
    pub inode: TrackedInode,
    pub removed: AtomicBool,
}

Represents a directory, which contains multiple directory entries.

This structure provides access to the directory’s inode, which stores information about the directory’s metadata and contents.

Fields
ffs: Weak<FastFileSystemInner>

Weak reference to the file system reference.

inode: TrackedInode

The inode associated with this directory.

removed: AtomicBool

Whether the directory is removed,

Implementations

impl Directory

pub fn new(inode: TrackedInode, ffs: Weak<FastFileSystemInner>) -> Option<Self>

Creates a new Directory from the given inode.

Arguments

inode: The tracked inode representing this directory.

ffs: A weak reference to the owning FastFileSystemInner.

Returns

Some(Directory): if the provided inode is valid and represents a directory.

None: if the inode is not of type File::Directory.

pub fn read_dir( &self, ffs: &FastFileSystemInner, ) -> Result<Vec<(InodeNumber, String)>, KernelError>

Reads the contents of the directory.

This function lists all the entries within the directory as a [Vec].

A single entry is a tuple that consists of inode number and file name, that is (InodeNumber, String).

Returns

Ok(()): If the directory was successfully read.

Err(Error): An error if the read operation fails.

pub fn find( &self, ffs: &FastFileSystemInner, entry: &str, ) -> Result<InodeNumber, KernelError>

Finds the inode number corresponding to a directory entry by name.

Arguments

ffs: Reference to the file system’s internal structure.

entry: The name of the directory entry to search for.

Returns

Ok(inode_number): if the entry is found in the directory.

Err(KernelError): if the entry is not found or other errors occurs.

pub fn add_entry( &self, ffs: &Arc<FastFileSystemInner>, entry: &str, ino: InodeNumber, tx: &RunningTransaction<'_>, ) -> Result<(), KernelError>

Adds a new entry to the directory.

Arguments

ffs: Reference to the file system’s internal structure.

entry: The name of the new directory entry.

is_dir: Whether the entry is a subdirectory (true) or a regular file (false).

tx: A running transaction used to persist metadata changes.

Returns

Ok(()): if the entry is successfully added.

Err(KernelError): if an error occurs (e.g., entry already exists or out of space).

pub fn take_entry( &self, ffs: &Arc<FastFileSystemInner>, entry: &str, tx: &RunningTransaction<'_>, ) -> Result<TrackedInode, KernelError>

Takes an existing entry out of the directory.

Arguments

ffs: Reference to the file system’s internal structure.

entry: The name of the entry to remove.

tx: A running transaction used to persist metadata changes.

Returns

Ok(()): if the entry is successfully removed.

Err(KernelError): if the entry does not exist or an I/O error occurs.