keos_project5::ffs

Struct FastFileSystem

pub struct FastFileSystem(pub Arc<FastFileSystemInner>);

A reference-counted wrapper around FastFileSystemInner.

This structure provides access to a Fast File System instance while ensuring safe concurrent access through an [Arc].

Tuple Fields
0: Arc<FastFileSystemInner>
Implementations

impl FastFileSystem

pub const ROOT_INODE_NUMBER: InodeNumber

The inode number of the root directory (/).

pub fn from_disk( disk: Disk, debug_journal: bool, disable_journal: bool, ) -> Result<Self, KernelError>

Loads a Fast File System from a given disk.

This function attempts to read the superblock from the disk. If the disk contains a valid FFS superblock, it returns an initialized FastFileSystem instance; otherwise, it returns None.

Parameters

disk: The disk device containing the filesystem.

Returns

Some(Self): If the filesystem is successfully loaded.

None: If the disk does not contain a valid Fast File System.

pub fn get_inode(&self, ino: InodeNumber) -> Result<TrackedInode, KernelError>

Retrieves an in-memory representation of the inode identified by ino.

This function looks up the inode in the Fast File System and returns a TrackedInode.